Prepare project for Railway deployment

+++ app/Http/Controllers/ProfileController.php
@@ -33,8 +33,13 @@ class ProfileController extends Controller
             $user->email_verified_at = null;
         }
 
-        if ($request->hasFile('avatar')) {
-            if ($user->avatar) {
+        if ($request->filled('avatar_url')) {
+            if ($user->avatar && !str_starts_with($user->avatar, 'http')) {
+                \Storage::disk('public')->delete($user->avatar);
+            }
+            $user->avatar = $request->input('avatar_url');
+        } elseif ($request->hasFile('avatar')) {
+            if ($user->avatar && !str_starts_with($user->avatar, 'http')) {
                 \Storage::disk('public')->delete($user->avatar);
             }
             $path = $request->file('avatar')->store('avatars', 'public');
